Move d-bus adaptors away from public headers
