+++ libconnman-qt/counter.cpp
@@ -15,6 +15,25 @@
 #include "counter.h"
 #include "networkmanager.h"
 
+class CounterAdaptor : public QDBusAbstractAdaptor
+{
+    Q_OBJECT
+    Q_CLASSINFO("D-Bus Interface", "net.connman.Counter")
+
+public:
+    explicit CounterAdaptor(Counter *parent);
+    virtual ~CounterAdaptor();
+
+public Q_SLOTS:
+    void Release();
+    void Usage(const QDBusObjectPath &service_path,
+                                const QVariantMap &home,
+                                const QVariantMap &roaming);
+
+private:
+    Counter *m_counter;
+};
+
 class CounterPrivate
 {
 public:
@@ -285,3 +304,4 @@ void CounterAdaptor::Usage(const QDBusObjectPath &service_path,
     }
 }
 
+#include "counter.moc"
