+++ libconnman-qt/sessionagent.cpp
@@ -23,6 +23,22 @@ Example:
 
   */
 
+class SessionNotificationAdaptor : public QDBusAbstractAdaptor
+{
+    Q_OBJECT
+    Q_CLASSINFO("D-Bus Interface", "net.connman.Notification")
+
+public:
+    explicit SessionNotificationAdaptor(SessionAgent* parent);
+    virtual ~SessionNotificationAdaptor();
+
+public Q_SLOTS:
+    void Release();
+    void Update(const QVariantMap &settings);
+private:
+    SessionAgent* m_sessionAgent;
+};
+
 class SessionAgentPrivate
 {
 public:
@@ -160,3 +176,5 @@ void SessionNotificationAdaptor::Update(const QVariantMap &settings)
 {
     m_sessionAgent->update(settings);
 }
+
+#include "sessionagent.moc"
