+++ libconnman-qt/useragent.cpp
@@ -12,6 +12,32 @@
 
 static const char AGENT_PATH[] = "/ConnectivityUserAgent";
 
+class AgentAdaptor : public QDBusAbstractAdaptor
+{
+    Q_OBJECT
+    Q_CLASSINFO("D-Bus Interface", "net.connman.Agent")
+
+public:
+    explicit AgentAdaptor(UserAgent* parent);
+    virtual ~AgentAdaptor();
+
+public Q_SLOTS:
+    void Release();
+    void ReportError(const QDBusObjectPath &service_path, const QString &error);
+    Q_NOREPLY void RequestBrowser(const QDBusObjectPath &service_path, const QString &url,
+                        const QDBusMessage &message);
+    void RequestConnect(const QDBusMessage &message);
+
+    Q_NOREPLY void RequestInput(const QDBusObjectPath &service_path,
+                                const QVariantMap &fields,
+                                const QDBusMessage &message);
+    void Cancel();
+
+private:
+    UserAgent* m_userAgent;
+    QElapsedTimer browserRequestTimer;
+};
+
 class UserAgentPrivate
 {
 public:
@@ -258,4 +284,4 @@ void AgentAdaptor::RequestConnect(const QDBusMessage &message)
     m_userAgent->requestConnect(message);
 }
 
-
+#include "useragent.moc"
